Added link to WIP Adobe XD prototype
By default, the prototype is in Full screen mode. If you press "Esc", you will be able to click the name of the project and a grid box in the top left corner. This will allow you to see all the different screens (there are 9 total). Try clicking around the different pages (hint: if you click something that is not linked to another page, the items that are linked to other pages will briefly flash). I will likely push images of each page later.